+++ history.md
@@ -1,9 +1,14 @@
 [2021/11/15]
+- Modify
+  + open.commons.spring.elastic.configurations.elasticsearch
+    + EndpointBuilderConfiguration.java
+    + SccBuilderConfiguration.java
+    + TccBuilderConfiguration.java  
 - New
   + open.commons.spring.elastic.configurations.elasticsearch
     + DefaultClientConfiguration.java
     + ElasticsearchClientConfiguration.java
-    + EndpointConfiguration.java
-    + SecureClientConfiguration.java
-    + TerminalClientConfiguration.java
+    + <strike>EndpointConfiguration.java</strike>
+    + <strike>SecureClientConfiguration.java</strike>
+    + <strike>TerminalClientConfiguration.java</strike>